pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-groq>=1.1.1",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.5",
    "mcp>=1.25.0",
    "google-genai>=0.2.0",
    "python-dotenv>=1.0.0",
]
